add pages to fetch all incident reports
