+++ lib/screens/super_admin/pages/incident_reports.dart
@@ -1,0 +1,228 @@
+import 'package:flutter/material.dart';
+import 'package:firebase_database/firebase_database.dart';
+
+class IncidentReportsScreen extends StatefulWidget {
+  const IncidentReportsScreen({super.key});
+  @override
+  _IncidentReportsScreenState createState() => _IncidentReportsScreenState();
+}
+
+class _IncidentReportsScreenState extends State<IncidentReportsScreen> {
+  final DatabaseReference dbRef = FirebaseDatabase.instance.ref();
+  List<Map<String, dynamic>> _reports = [];
+  bool _loading = true;
+
+  @override
+  void initState() {
+    super.initState();
+    _fetchReports();
+  }
+
+  Future<void> _fetchReports() async {
+    try {
+      final snapshot = await dbRef.child('incident-reports').get();
+      if (snapshot.exists) {
+        List<Map<String, dynamic>> reports = [];
+        snapshot.children.forEach((reportSnapshot) {
+          final report = Map<String, dynamic>.from(reportSnapshot.value as Map);
+          reports.add(report);
+        });
+
+        setState(() {
+          _reports = reports;
+          _loading = false;
+        });
+      } else {
+        setState(() {
+          _reports = [];
+          _loading = false;
+        });
+      }
+    } catch (e) {
+      print("Error fetching reports: $e");
+      setState(() {
+        _loading = false;
+      });
+    }
+  }
+
+  void _showReportDetailsModal(
+      BuildContext context, Map<String, dynamic> report) {
+    showDialog(
+      context: context,
+      barrierDismissible: true,
+      builder: (BuildContext context) {
+        return Dialog(
+          shape:
+              RoundedRectangleBorder(borderRadius: BorderRadius.circular(10.0)),
+          child: Stack(
+            children: [
+              Padding(
+                padding: const EdgeInsets.all(10.0),
+                child: SingleChildScrollView(
+                    child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.end,
+                      children: [
+                        IconButton(
+                          icon: const Icon(Icons.close),
+                          iconSize: 20,
+                          onPressed: () {
+                            Navigator.of(context).pop();
+                          },
+                        ),
+                      ],
+                    ),
+                    report['image_url'] != null
+                        ? Image.network(
+                            report['image_url'],
+                            height: 200,
+                            width: double.infinity,
+                            fit: BoxFit.cover,
+                          )
+                        : const Icon(Icons.image_not_supported, size: 100),
+                    const SizedBox(height: 20),
+                    Text('${report['details']}',
+                        style: const TextStyle(fontSize: 14)),
+                    const SizedBox(height: 8),
+                    Text('${report['reporter_name']}',
+                        style: const TextStyle(fontSize: 12)),
+                    Text('${report['reporter_email']}',
+                        style: const TextStyle(fontSize: 12)),
+                    Text('${report['address']}',
+                        style: const TextStyle(fontSize: 12)),
+                    Text('${report['timestamp']}',
+                        style: const TextStyle(fontSize: 12)),
+                  ],
+                )),
+              ),
+            ],
+          ),
+        );
+      },
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Incident Reports'),
+      ),
+      body: _loading
+          ? const Center(
+              child: CircularProgressIndicator(),
+            )
+          : _reports.isEmpty
+              ? const Center(child: Text('No reports found.'))
+              : ListView.builder(
+                  itemCount: _reports.length,
+                  itemBuilder: (context, index) {
+                    final report = _reports[index];
+                    final String detailsSnippet = report['details'] != null
+                        ? (report['details'].length > 50
+                            ? report['details'].substring(0, 50) + '...'
+                            : report['details'])
+                        : 'No Details';
+                    final String addressSnippet = report['address'] != null
+                        ? (report['address'].length > 60
+                            ? report['address'].substring(0, 60) + '...'
+                            : report['address'])
+                        : 'No Address';
+                    return GestureDetector(
+                      onTap: () {
+                        _showReportDetailsModal(context, report);
+                      },
+                      child: SizedBox(
+                        height: 200,
+                        width: double.infinity,
+                        child: Card(
+                          elevation: 4.0,
+                          margin: const EdgeInsets.all(10.0),
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Row(
+                                mainAxisAlignment: MainAxisAlignment.end,
+                                children: [
+                                  Padding(
+                                    padding: const EdgeInsets.symmetric(
+                                        vertical: 5.0, horizontal: 5.0),
+                                    child: Container(
+                                      padding: const EdgeInsets.symmetric(
+                                          horizontal: 10, vertical: 5),
+                                      decoration: BoxDecoration(
+                                        color: Colors.red,
+                                        borderRadius:
+                                            BorderRadius.circular(100),
+                                      ),
+                                      child: Text('${report['status']}',
+                                          style: const TextStyle(
+                                              fontSize: 10,
+                                              color: Colors.white)),
+                                    ),
+                                  ),
+                                ],
+                              ),
+                              Row(
+                                crossAxisAlignment: CrossAxisAlignment.center,
+                                children: [
+                                  Padding(
+                                    padding: const EdgeInsets.all(10.0),
+                                    child: report['image_url'] != null
+                                        ? Image.network(
+                                            report['image_url'],
+                                            width: 50,
+                                            height: 50,
+                                            fit: BoxFit.cover,
+                                          )
+                                        : const Icon(Icons.image_not_supported,
+                                            size: 50),
+                                  ),
+                                  Expanded(
+                                    child: Padding(
+                                      padding: const EdgeInsets.symmetric(
+                                          vertical: 20.0, horizontal: 20.0),
+                                      child: Column(
+                                        crossAxisAlignment:
+                                            CrossAxisAlignment.start,
+                                        mainAxisAlignment:
+                                            MainAxisAlignment.center,
+                                        children: [
+                                          Text(
+                                            detailsSnippet,
+                                            style: const TextStyle(
+                                              fontSize: 15,
+                                              fontWeight: FontWeight.w400,
+                                              height: 1,
+                                            ),
+                                          ),
+                                          const SizedBox(height: 8),
+                                          Text(
+                                            addressSnippet,
+                                            style:
+                                                const TextStyle(fontSize: 12),
+                                          ),
+                                          Text(
+                                            '${report['timestamp']}',
+                                            style:
+                                                const TextStyle(fontSize: 12),
+                                          ),
+                                        ],
+                                      ),
+                                    ),
+                                  ),
+                                ],
+                              ),
+                            ],
+                          ),
+                        ),
+                      ),
+                    );
+                  },
+                ),
+    );
+  }
+}
